+++ src/plugins/intel_gpu/src/kernel_selector/kernel_selector.cpp
@@ -89,7 +89,10 @@ KernelsData kernel_selector_base::GetNaiveBestKernel(const KernelList& all_impls
                 if (!params.is_shape_agnostic) {
                     for (size_t k = 0; k < kds[0].kernels.size(); ++k) {
                         auto gws = kds[0].kernels[k].params.workGroups.global;
-                        kernelsData[0].kernels[k].skip_execution = (std::accumulate(gws.begin(), gws.end(), 1, std::multiplies<size_t>()) == 0);
+                        kernelsData[0].kernels[k].skip_execution = (std::accumulate(gws.begin(),
+                                                                                    gws.end(),
+                                                                                    static_cast<size_t>(1),
+                                                                                    std::multiplies<size_t>()) == 0);
                     }
                 }
                 break;
