+++ src/plugins/intel_gpu/tests/shape_infer/crop_si_test.cpp
@@ -61,7 +61,7 @@ TEST_P(crop_si_test, shape_infer) {
 
     for (size_t output_idx = 0; output_idx < p.expected_layouts.size(); output_idx++) {
         auto prim_id = "crop.out" + std::to_string(output_idx);
-        auto crop_prim = std::make_shared<crop>(prim_id, input_prim_ids, p.reference_input_size, p.offsets[output_idx], op_mode, output_idx, p.param_num_splits);
+        auto crop_prim = std::make_shared<crop>(prim_id, input_prim_ids, p.reference_input_size, p.offsets[output_idx], op_mode, static_cast<int>(output_idx), p.param_num_splits);
         auto& crop_node = prog.get_or_create(crop_prim);
 
         for (auto& prim : input_prims) {
